feat(sendTargetedEmails): Move getting email templates functionality into separate file
